fix: 修复 pg_catalog 特殊语法前缀、测试种子与 nanoid 高危漏洞
- 新增迁移 20260816000007：重建 search_workbench_v2 / suggest_ledger_recurrences /
  apply_workbench_command_v2，去掉非法的 pg_catalog.least/greatest/coalesce/trim 前缀
  （这些是语法层特殊表达式，带前缀调用在 PL/pgSQL 创建时不报错、运行时抛 42883，
  云端三个函数此前处于不可用状态）
- recurrence_edges.sql：month_day 字面量 31 显式转 smallint（int4→int2 仅为赋值转换，
  函数参数解析失败导致 function does not exist）
- rls_isolation.sql：workout_exercises 种子插入显式提供 user_id（该列自
  20260813000001 起 NOT NULL DEFAULT auth.uid()，超级用户会话无认证上下文）
- 重新生成损坏的 database.types.ts 类型基线（原文件第 20000 字符处截断），eslint 忽略生成目录
- npm audit fix：nanoid 升至 3.3.18，消除 1 个 high 级漏洞
- CI 严格化：fetch-depth 0 + MIGRATION_BASE_REF 基线对比 + 类型漂移强制检查 +
  迁移 append-only 防篡改校验（git 基线 SHA 对比）

+++ supabase/tests/database/recurrence_edges.sql
@@ -7,19 +7,19 @@ select extensions.plan(20);
 
 -- ---------- 匹配算法：monthly month_day=31 压尾 ----------
 select extensions.is(
-  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31,'2026-01-31','2027-02-28'),
+  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31::smallint,'2026-01-31','2027-02-28'),
   true,'monthly day-31 clamps to Feb 28 in a common year');
 select extensions.is(
-  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31,'2026-01-31','2028-02-29'),
+  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31::smallint,'2026-01-31','2028-02-29'),
   true,'monthly day-31 clamps to Feb 29 in a leap year');
 select extensions.is(
-  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31,'2026-01-31','2026-04-30'),
+  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31::smallint,'2026-01-31','2026-04-30'),
   true,'monthly day-31 clamps to Apr 30');
 select extensions.is(
-  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31,'2026-01-31','2026-03-31'),
+  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31::smallint,'2026-01-31','2026-03-31'),
   true,'monthly day-31 matches a real 31st');
 select extensions.is(
-  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31,'2026-01-31','2026-03-30'),
+  public.recurrence_occurrence_matches('monthly',1,'{}'::smallint[],31::smallint,'2026-01-31','2026-03-30'),
   false,'monthly day-31 does not match the 30th of a 31-day month');
 
 -- ---------- 匹配算法：yearly 2/29 规则（闰年与平年压尾） ----------
